simplify the elaborator logic. - make sure all the concats of theta type after elaboration are merged so that no duplicate thetas exist - make sure we don't generate `@` constraints for tycons that have a constraint return kind, even tctycon
Becuase we never reduce any well formed constraints. we get "extra" monomorphic constraints such as `IO @ Int` etc.
We need to turn on the `NoMonoMorphicRestriction` in such cases. This should be temp thing, and ideally I would like to get
rid of these monomorphic restrictions, but may be it is not possible in general

The only place where we eagerly reduce `@` constraints is when we have to check a deriving clause.
eg. in `deriving (C t) via (D a)`. We don't really want to keep around `D @ a => D a` as this messes the representation type.