- This commit executes compilation for stage2:exe:ghc-bin   This does not however generate a proper stage2 binary as we have changed the logic of when wf type families reduce. This is a good start however towards a stable branch. - Specifically wf type families reduces when   1. The type is a SPEC pragma   2. A Derive context clause - They do not reduce in   1. top level bindings
The reason to do so is to make boot files happy and get rid of `wf_free_type` function.
Fixes #25.
There are some places in the desugaring pass were we now have to provide a bogus `unitExpr` as
the original type of the function expects a `T @ a` constraint. I have not yet figured out
all the places where we need this extra argument yet and hence it does not generate a proper binary.
cf. when desugaring an expression of the form `x::Float; x = 0.4`, `dsFractionalListToRational` is called
that uses `mkRationalBase10Name::Rational -> Integer -> Rational` which elaborates to
`Ratio @ Integer => Rational -> Integer -> Rational`. And as it is a top level bindings we need to pass in an extra unitExpr.