Have found root of problem w/ Closed TF exporting, but...
will need time to fix this properly. Issue is:
- WF TF mirror is imported correctly
- However, WF TF parent _recreates_ its WF TF child. Only problem is
  the parent does not have its child's axioms! So it gives it its own.

A bad fix would be to force a re-lookup of the wf child at some point after we
know the wf child has been made. But this is hacky. Will want to treat this like
associated types: add some data to the IFaceFamily branch of IFaceDecl data
type, like the child's axioms (or even the child itself, if possible).