Elaborate closed TFs from boot files
We have to add a lot of constraints like:
  WFT (SyntaxExprGhc 'Renamed)
because closed TF applications are not elaborated in constraint lists,
as we thought closed TFs could not have their equations in boot files.
However, you can have them -- just not open TF instances. So things like
the above will properly reduce to just () :: Constraint.